feat: `BetaFromMeanVar` and spin models reorganized (#394)
feat: `BetaFromMeanVar` and reorganize spin models and update documentation
