+++ src/gwkokab/models/_models.py
@@ -29,10 +29,8 @@ from numpyro.distributions import (
     Distribution,
     DoublyTruncatedPowerLaw,
     MixtureGeneral,
-    MultivariateNormal,
     Normal,
     TruncatedNormal,
-    Uniform,
 )
 from numpyro.distributions.util import promote_shapes, validate_sample
 
@@ -47,8 +45,6 @@ from .utils import (
 
 __all__ = [
     "FlexibleMixtureModel",
-    "GaussianSpinModel",
-    "IndependentSpinOrientationGaussianIsotropic",
     "MassGapModel",
     "NDistribution",
     "PowerlawPrimaryMassRatio",
@@ -56,94 +52,6 @@ __all__ = [
     "SmoothedPowerlawPrimaryMassRatio",
     "Wysocki2019MassModel",
 ]
-
-
-def GaussianSpinModel(
-    mu_eff, sigma_eff, mu_p, sigma_p, rho, *, validate_args=None
-) -> MultivariateNormal:
-    r"""Bivariate normal distribution for the effective and precessing spins.
-    See Eq. (D3) and (D4) in `Population Properties of Compact Objects from
-    the Second LIGO-Virgo Gravitational-Wave Transient
-    Catalog <https://arxiv.org/abs/2010.14533>`_.
-
-    .. math::
-        \left(\chi_{\text{eff}}, \chi_{p}\right) \sim \mathcal{N}\left(
-            \begin{bmatrix}
-                \mu_{\text{eff}} \\ \mu_{p}
-            \end{bmatrix},
-            \begin{bmatrix}
-                \sigma_{\text{eff}}^2 & \rho \sigma_{\text{eff}} \sigma_{p} \\
-                \rho \sigma_{\text{eff}} \sigma_{p} & \sigma_{p}^2
-            \end{bmatrix}
-        \right)
-
-    where :math:`\chi_{\text{eff}}` is the effective spin and
-    :math:`\chi_{\text{eff}}\in[-1,1]` and :math:`\chi_{p}` is the precessing spin and
-    :math:`\chi_{p}\in[0,1]`.
-
-    :param mu_eff: mean of the effective spin
-    :param sigma_eff: standard deviation of the effective spin
-    :param mu_p: mean of the precessing spin
-    :param sigma_p: standard deviation of the precessing spin
-    :param rho: correlation coefficient between the effective and precessing
-        spins
-    :return: Multivariate normal distribution for the effective and precessing
-        spins
-    """
-    return MultivariateNormal(
-        loc=jnp.array([mu_eff, mu_p]),
-        covariance_matrix=jnp.array(
-            [
-                [
-                    jnp.square(sigma_eff),
-                    jnp.multiply(rho, jnp.multiply(sigma_eff, sigma_p)),
-                ],
-                [
-                    jnp.multiply(rho, jnp.multiply(sigma_eff, sigma_p)),
-                    jnp.square(sigma_p),
-                ],
-            ]
-        ),
-        validate_args=validate_args,
-    )
-
-
-def IndependentSpinOrientationGaussianIsotropic(
-    zeta, sigma1, sigma2, *, validate_args=None
-) -> MixtureGeneral:
-    r"""A mixture model of spin orientations with isotropic and normally
-    distributed components. See Eq. (4) of `Determining the population
-    properties of spinning black holes <https://arxiv.org/abs/1704.08370>`_.
-
-    .. math::
-        p(z_1,z_2\mid\zeta,\sigma_1,\sigma_2) = \frac{1-\zeta}{4} +
-        \zeta\mathbb{I}_{[-1,1]}(z_1)\mathbb{I}_{[-1,1]}(z_2)
-        \mathcal{N}(z_1\mid 1,\sigma_1)\mathcal{N}(z_2\mid 1,\sigma_2)
-
-    where :math:`\mathbb{I}(\cdot)` is the indicator function.
-
-    :param zeta: The mixing probability of the second component.
-    :param sigma1: The standard deviation of the first component.
-    :param sigma2: The standard deviation of the second component.
-    :return: Mixture model of spin orientations.
-    """
-    mixing_probs = jnp.array([1 - zeta, zeta])
-    component_0_dist = Uniform(low=-1, high=1, validate_args=validate_args)
-    component_1_dist = TruncatedNormal(
-        loc=1.0,
-        scale=jnp.array([sigma1, sigma2]),
-        low=-1,
-        high=1,
-        validate_args=validate_args,
-    )
-    return MixtureGeneral(
-        mixing_distribution=CategoricalProbs(
-            probs=mixing_probs, validate_args=validate_args
-        ),
-        component_distributions=[component_0_dist, component_1_dist],
-        support=constraints.real,
-        validate_args=validate_args,
-    )
 
 
 def NDistribution(
